Extract `LegacyInteractionBarItem` to new file
