+++ Mlem/App/Enums/Interaction/InteractionBarConfiguration.swift
@@ -12,31 +12,8 @@ import MlemMiddleware
 import SwiftUI
 
 protocol InteractionBarConfiguration {
-    associatedtype ActionType: ActionTypeProviding
-    
-    typealias Item = InteractionConfigurationItem<ActionType>
-
-    static var availableActions: ActionSeedSections { get }
-    
-    init()
-}
-
-enum InteractionConfigurationItem<ActionType: ActionTypeProviding>: Codable, Hashable {
-    case action(ActionType)
-    case counter(CounterType)
-    
-    func toInteractionBarItem() -> InteractionBarItem {
-        switch self {
-        case let .action(action):
-            .action(action.actionSeed)
-        case let .counter(counter):
-            .counter(counter)
-        }
-    }
-}
-
-protocol ActionTypeProviding: Codable, Hashable {
-    var actionSeed: ActionSeed { get }
+    associatedtype ActionType: LegacyActionTypeProviding
+    typealias Item = LegacyInteractionBarItem<ActionType>
 }
 
 struct MockReadoutAppearance {
